% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archipelago.R
\name{archipelago_plot}
\alias{archipelago_plot}
\title{Archipelago Plot}
\usage{
archipelago_plot(
  df1,
  df2,
  plot_title = "Archipelago Plot",
  add_title = FALSE,
  plot_subtitle =
    "Variant Set Association Test\\nwith individual variant contributions\\nand contributions for significant VSAT",
  add_subtitle = FALSE,
  chr_ticks = TRUE,
  show_legend = TRUE,
  color_theme = NULL,
  custom_colors = NULL,
  color_labels = c("Chr: Individual p-val", "Chr: Individual p-val",
    "Individual variant\\nfrom enriched VSAT", "VSAT p-val"),
  crit_val = NULL,
  point_size = 1,
  output_path = "archipelago_plot.pdf",
  output_raw = "vsat_raw_plot.pdf"
)
}
\arguments{
\item{df1}{A dataframe of variant-set data.}

\item{df2}{A dataframe of variant data.}

\item{plot_title}{Title for the plot.}

\item{add_title}{Logical; add title if TRUE.}

\item{plot_subtitle}{Subtitle for the plot.}

\item{add_subtitle}{Logical; add subtitle if TRUE.}

\item{chr_ticks}{Logical; show chromosome ticks if TRUE.}

\item{show_legend}{Logical; display the legend if TRUE.}

\item{color_theme}{Name of the colour theme.}

\item{custom_colors}{Vector of custom colours.}

\item{color_labels}{Labels for the colour groups.}

\item{crit_val}{Critical p-value threshold.}

\item{point_size}{Size of the points.}

\item{output_path}{File path for the plot.}

\item{output_raw}{File path for the raw plot.}
}
\value{
A ggplot object.
}
\description{
Plot p-values from variant and variant-set tests.
}
\examples{
\dontrun{
  # Load test data for df1 and df2
  data("vsat_pval")
  data("variant_pval")

  # Use default settings
  archipelago_plot(vsat_pval, variant_pval)

  # Use the 'alice' colour theme
  archipelago_plot(vsat_pval, variant_pval, color_theme = "alice")

  # Custom everything
  output_path <- "./archipelago_plot_custom_color.pdf"
  output_raw <- "./vsat_raw_plot.pdf"
  color_labels <- c("Label_1", "Label_2", "Label_3", "Label_4")
  custom_colors <- c("#9abfd8", "#cac1f3", "#371c4b", "#2a5b7f")
  plot_title <- "Title"
  plot_subtitle <- "Subtitle"
  crit_val <- .05/300
  point_size <- .5

  archipelago_plot(
    df1 = vsat_pval,
    df2 = variant_pval,
    add_title = TRUE,
    plot_title = plot_title,
    add_subtitle = TRUE,
    plot_subtitle = plot_subtitle,
    show_legend = TRUE,
    crit_val = crit_val,
    point_size = point_size,
    chr_ticks = FALSE,
    show_legend = TRUE,
    custom_colors = custom_colors,
    color_labels = color_labels,
    output_path = output_path,
    output_raw = output_raw
  )
}
}
