% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variant_pval}
\alias{variant_pval}
\title{Example variant data}
\format{
A data frame with columns:
\itemize{
\item set_ID
\item BP
\item P
\item CHR
\item SNP
}
}
\usage{
variant_pval
}
\description{
Example variant-level p-values from a genome-wide association study.
}
\keyword{datasets}
